use float for pointDelta for half points in rubric

+++ cpAddRubric.py
@@ -73,7 +73,7 @@ rubric file should contain something like this
                     state = "category"
                 else:
                     spacePos = line.find(" ")
-                    pointDelta = int(line[:spacePos])
+                    pointDelta = float(line[:spacePos])
                     text = line[spacePos+1:]
                     if not rubricCategory.hasRubricComment(text, pointDelta):
                         print(f"add comment: {text} : {pointDelta}")
